Fix data race in use of cmd output buffers.
The bytes.Buffers used to store the output of subprocesses are managed with a
sync.Pool. However, before this change, they were being returned to the Pool
while the slice underlying the Buffer was still in use by caller methods. If
that Buffer was then used by another caller, reads and writes to the underlying
slice had the potential race with one another (which is detectable by running
with the "-race" flag enabled in a binary and/or a test).

This change puts the responsibility on the caller of cmdOutput to put the buffer
back in the pool once it is done reading data from it, ensuring no two methods
use its slice at the same time.

Signed-off-by: Erik Sipsma <sipsma@amazon.com>